templatizing

+++ src/data/PartialFov.h
@@ -14,16 +14,16 @@ public:
     PartialFov() {}
 
 
-    PartialFov(const std::string &path, uint32_t x, uint32_t y) : path(path), x(x), y(y) {}
+    PartialFov(const std::string &path, size_t x, size_t y) : path(path), x(x), y(y) {}
 
-    PartialFov(const std::string &path, uint32_t x, uint32_t y, const cv::Rect &tileOverlap) : path(path), x(x), y(y),
+    PartialFov(const std::string &path, size_t x, size_t y, const cv::Rect &tileOverlap) : path(path), x(x), y(y),
                                                                                         tileOverlap(tileOverlap) {}
 
-    PartialFov(const std::string &path, uint32_t x, uint32_t y, const cv::Rect &tileOverlap,
+    PartialFov(const std::string &path, size_t x, size_t y, const cv::Rect &tileOverlap,
         const cv::Rect &globalCoordinatesTileOverlap) : path(path), x(x), y(y), tileOverlap(tileOverlap),
                                                         globalCoordinatesTileOverlap(globalCoordinatesTileOverlap) {}
 
-    PartialFov(const std::string &path, uint32_t x, uint32_t y, const cv::Rect &tileOverlap,
+    PartialFov(const std::string &path, size_t x, size_t y, const cv::Rect &tileOverlap,
         const cv::Rect &globalCoordinatesTileOverlap, const cv::Rect &FovCoordOverlap) : path(path), x(x), y(y),
                                                                                          tileOverlap(tileOverlap),
                                                                                          globalCoordinatesTileOverlap(
@@ -36,11 +36,11 @@ public:
         return path;
     }
 
-    uint32_t getX() const {
+    size_t getX() const {
         return x;
     }
 
-    uint32_t getY() const {
+    size_t getY() const {
         return y;
     }
 
@@ -59,8 +59,8 @@ public:
 private:
 
     std::string path;
-    uint32_t x;
-    uint32_t y;
+    size_t x;
+    size_t y;
     cv::Rect tileOverlap;
     cv::Rect globalCoordinatesTileOverlap;
     cv::Rect FovCoordOverlap;
